Adds first work at dither, with build support.
The build command is based on an older build.py script from my dotfiles repo (it was called "dotfiles_jinja", but this is likely to have changed since writing).

This is very rough and not ready for use yet.  This was mostly committed to get the work off my laptop, but I hope you'll forgive that of the first commit. :)