Changes dependencies on "os.path" to "os:path".
The full resource name was '.org.python.stdlib.os.path', and is now '.org.python.stdlib.os:path'. It describes a dependency on the os.path sub-module from the Python standard library.

For resources used from the Python standard library, the planned convention is that everything between 'org.python.stdlib.' and the ':' is what should be imported, and everything after the ':' is what should be accessed on the imported module.

For this reason, ".org.python.stdlib.os:path" is now used because according to the os.path docs, the correct way to use os.path is to import the os module, then access the path sub-module on os.