build.core: Refactor unit tests to move common setup into fixtures.
Some of this is enabled by fang v0.7.0 supporting the allow_override keyword arg to ResourceProvider.register*() methods. This allows having a setup of registering several providers common to a set of tests, with some tests in the set overriding providers for specific resources as needed (eg to add a test recorder on the fake given for one resource).